# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
...
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"Required_any_": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			},
...
			"ICommunications": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"session",
							"agreements",
							"disputes",
							"users"
						]
					},
					"referenceID": {
						"type": "string"
					},
					"partiesUIDs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"lastUpdate": {
						"type": "string",
						"format": "date-time"
					},
					"messages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"deliveryDetails": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"referenceType",
					"referenceID",
					"partiesUIDs",
					"lastUpdate",
					"messages",
					"deliveryDetails"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ICommunications_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"session",
							"agreements",
							"disputes",
							"users"
						]
					},
					"referenceID": {
						"type": "string"
					},
					"partiesUIDs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"lastUpdate": {
						"type": "string",
						"format": "date-time"
					},
					"messages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"deliveryDetails": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IDispute": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionID": {
						"type": "string"
					},
					"ﬁrstPartyUID": {
						"type": "string"
					},
					"secondUID": {
						"type": "string"
					},
					"topic": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"sent",
							"received",
							"in-progress",
							"suspended",
							"rejected",
							"resolved"
						]
					},
					"resolverUID": {
						"type": "string"
					},
					"inProgressDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"suspendedDate": {
						"type": "string",
						"format": "date-time"
					},
					"closedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"sessionID",
					"ﬁrstPartyUID",
					"secondUID",
					"topic",
					"details",
					"attachments",
					"status",
					"resolverUID",
					"inProgressDate",
					"receivedDate",
					"suspendedDate",
					"closedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IDispute_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionID": {
						"type": "string"
					},
					"ﬁrstPartyUID": {
						"type": "string"
					},
					"secondUID": {
						"type": "string"
					},
					"topic": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"sent",
							"received",
							"in-progress",
							"suspended",
							"rejected",
							"resolved"
						]
					},
					"resolverUID": {
						"type": "string"
					},
					"inProgressDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"suspendedDate": {
						"type": "string",
						"format": "date-time"
					},
					"closedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IEmloyees": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"uID": {
						"type": "string"
					},
					"roleID": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"attachments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"uID",
					"roleID",
					"salary",
					"attachments"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IEmloyees_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"uID": {
						"type": "string"
					},
					"roleID": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"attachments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "newproj",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/Communications/{communicationId}": {
			"get": {
				"operationId": "GetCommunication",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ICommunications"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested communication in not found"
					}
				},
				"description": "Get a communication details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "communicationId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			},
			"delete": {
				"operationId": "DeleteCommunication",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/ICommunications"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested communication in not found"
					}
				},
				"description": "Delete a communication",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "communicationId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			}
		},
		"/Communications/create": {
			"post": {
				"operationId": "CreateCommunication",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICommunications"
								},
								"examples": {
									"Example 1": {
										"value": {
											"referenceType": "agreements",
											"referenceID": "1",
											"partiesUIDs": [
												"",
												""
											],
											"lastUpdate": "2022-09-02T21:00:00.000Z",
											"messages": [
												""
											],
											"deliveryDetails": [
												""
											]
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a communication",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICommunications"
							}
						}
					}
				}
			}
		},
		"/Communications/update/{communicationId}": {
			"put": {
				"operationId": "UpdateCommunication",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ICommunications"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a communication",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "communicationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ICommunications_"
							}
						}
					}
				}
			}
		},
...
		"/Dispute/{disputeId}": {
			"get": {
				"operationId": "GetDispute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IDispute"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested dispute in not found"
					}
				},
				"description": "Get a dispute details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "disputeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			},
			"delete": {
				"operationId": "DeleteDispute",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/IDispute"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested dispute in not found"
					}
				},
				"description": "Delete a dispute",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "disputeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			}
		},
		"/Dispute/create": {
			"post": {
				"operationId": "CreateDispute",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IDispute"
								},
								"examples": {
									"Example 1": {
										"value": {
											"sessionID": "1",
											"ﬁrstPartyUID": "1",
											"secondUID": "1",
											"topic": "",
											"details": "",
											"attachments": [
												""
											],
											"status": "in-progress",
											"resolverUID": "1",
											"inProgressDate": "2022-09-02T21:00:00.000Z",
											"receivedDate": "2022-09-02T21:00:00.000Z",
											"suspendedDate": "2022-09-02T21:00:00.000Z",
											"closedDate": "2022-09-02T21:00:00.000Z"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a dispute",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDispute"
							}
						}
					}
				}
			}
		},
		"/Dispute/update/{disputeId}": {
			"put": {
				"operationId": "UpdateDispute",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IDispute"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a dispute",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "disputeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IDispute_"
							}
						}
					}
				}
			}
		},
...
		"/Emloyees/{emloyeId}": {
			"get": {
				"operationId": "GetEmloye",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEmloyees"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested emloyees in not found"
					}
				},
				"description": "Get a emloye details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "emloyeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			},
			"delete": {
				"operationId": "DeleteEmloye",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/IEmloyees"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested emloye in not found"
					}
				},
				"description": "Delete a emloye",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "emloyeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			}
		},
		"/Emloyees/create": {
			"post": {
				"operationId": "CreateEmloye",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEmloyees"
								},
								"examples": {
									"Example 1": {
										"value": {
											"uID": "1",
											"roleID": "1",
											"salary": 500,
											"attachments": [
												""
											]
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a emloye",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEmloyees"
							}
						}
					}
				}
			}
		},
		"/Emloyees/update/{emloyeId}": {
			"put": {
				"operationId": "UpdateEmloye",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEmloyees"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a emloye",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "emloyeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IEmloyees_"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}